backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
redis==5.0.1
docker==7.0.0
pypdf==4.0.1
python-docx==1.1.0
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
gTTS==2.5.1
